moved TechnologiesIcons to utils folder
